WFCORE-2083 Only register transformers when needed
- lazy load parsers
